Baby steps to get new Authentication Cookie happening. Still using old Cookies at this revision, but some groundwork is now laid for the full implementation.
Deprecate bb_random_pass() in favour of wp_generate_password()

Implement new wp_salt() with some minor changes. We can't try to read or write a secret during installation as there is no database. This was a problem brought about by using referrer checks on the forms, which use wp_hash() which in turn use wp_salt().

Some fixes to the installer. Selecting a WordPress admin from the dropdown to become a KeyMaster now works again.